+++ src/main/java/org/generationcp/middleware/service/api/phenotype/PhenotypeSearchDTO.java
@@ -1,5 +1,3 @@
-
-
 package org.generationcp.middleware.service.api.phenotype;
 
 import com.fasterxml.jackson.annotation.JsonView;
